requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.1.3
python-dateutil==2.9.0
